// src/types/pessoaTypes.ts
export type PessoaCreateType = {
    cpf: string;
    nome: string;
    sobrenome: string;
    data_nascimento: string;
    id_cidade: string;
    bairro: string;
    cep: string;
    logradouro: string;
    numero_residencial: string;
    ddd: string;
    telefone: string;
    email: string;
};

export type PessoaUpdateType = {
    cpf: string;
    nome: string;
    sobrenome: string;
    data_nascimento: string;
    id_cidade: string;
    bairro: string;
    cep: string;
    logradouro: string;
    numero_residencial: string;
    ddd: string;
    telefone: string;
    email: string;
};